Move `embedded-hal` 1.0 traits out of `hal-unproven` feature.
